Restructured cpp
