+++ norse/csrc/super.h
@@ -1,0 +1,7 @@
+#include <torch/torch.h>
+
+namespace norse
+{
+    torch::Tensor superfun(torch::Tensor input, torch::Tensor alpha);
+
+}
